requirements.txt
python-dotenv
requests
yt-dlp
youtube-search-python